+++ based/src/ray.rs
@@ -10,7 +10,7 @@ impl Ray {
         Ray { o, d }
     }
 
-    pub fn at(&self, t: f32) -> Vec3 {
+    pub fn at(&self, t: f64) -> Vec3 {
         self.o + t * self.d
     }
 }
